fix: address PR #336 review comments — XSS sanitization, deterministic IDs, slider data attrs, file splits
- Add wp_kses() sanitization to text buffer flush, table HTML, and code block content
- Add wp_kses_post() to REST input HTML before conversion
- Store uniqueId/formId as block attributes for accordion, form, and tabs (deterministic IDs matching save.js)
- Emit all data-* attributes, CSS custom properties, and modifier classes in slider handler
- Split class-dsgo-handlers.php into form/interactive/layout handler files (300-line guideline)
- Split html-converter-test.php into focused test files per handler category
- Deduplicate build_inner_content_for_blocks via public static method
- Guard has-text-color marker class in attribute mapper
- Fix code block test assertion for esc_html() output
